Add Watermark Reporting backend support
This adds the necessary migrations and changes to collect the data
necessary for watermark reporting. In particular it adds:

* host sockets
* host count total
* vm count total

to metrics and metric rollups.

It also adds a model for reporting on this data and the necessary copy
to integrate it into the UI.

It also adds the ability to group reports by month of performance timestamp.

Resolves https://trello.com/c/Ab3u7ZXR

(transferred from ManageIQ/manageiq@3f4df0c0077bfe803b583cd32e64b54f41ad561e)