Add a spec to be sure migrations will be run in order on upgrade
If two migrations are added which add columns on the same table,
and only the later one gets into a release, migrations to the next
release will have the columns in the reverse order when compared
with newly created appliances or databases migrated from scratch.

This spec tests that no migrations which are not in the list of
released migrations have a timestamp earlier than the latest
released migration. If this happens, the developer has to regenerate
the migration so that everything will run in the proper order.

This will help prevent us from getting into the situation which
caused https://bugzilla.redhat.com/show_bug.cgi?id=1331114

(transferred from ManageIQ/manageiq@392423a11a4f29c160414dae96c3c862a9f03c0d)