Fix sporadic failures in the replication test setup
Fetching the list of migrations from the git API was subject to
sporadic failures because of request limits.

This change fetches the darga branch locally and queries it to
determine the migrations, removing the need for an API call.

(transferred from ManageIQ/manageiq@3aa83193c6d4cfb842fde171ef02314219e213e1)