Only migrate rows in the current region
This was causing an issue with replication when rows which actually
belonged to a region were migrated.

The new rows in the join table got an id in the global region.

This caused replication to fail with a unique constraint error
when trying to replicate the new rows in the regional database.

https://bugzilla.redhat.com/show_bug.cgi?id=1361218

(transferred from ManageIQ/manageiq@6c0ab8b32f8f621474588bfe94b62fc25692bc0d)