Schema/models for tenancy sharing feature
This should allow us to:

As a single user
  share multiple "shareables" (aka, MiqTemplates, ServiceTemplates, anything...)
  contained within a single share
  having a set of miq_product_features for that share
  with multiple tenants

The shareables are contained within the user level object, a Share.

The Share has a set of miq_product_features, assumed to be a subset of the
user's current features allotted by their current miq_user_role.

A Share is connected to many shareables via a ShareMember (It's a
*member* of a share). Similarly, a Share is connected to many tenants
via a ShareReceiver (they are the things that *receive* the share).

The `shares` table also has `permission_mode` and
`block_tenant_inheritance` columns. These are *unused* for now, but
requested to be included in this schema change ahead of time. The
permission_mode is assumed to allow for a 'basic' version of selecting
miq_product_features (read, write, etc). block_tenant_inheritance is
assumed to allow for shares to only be seen at the direct tenant
assigned and not allowed to be seen from child tenants.

Lastly, note that MiqTemplate and ServiceTemplate do NOT have an
ActiveRecord association defined in their models. It would be nice to
have MiqTemplate.share_members but as MiqTemplate is a VmOrTemplate,
Rails uses this as the shareable_type to be looked up in share_members
(which clearly is not the correct thing, and returns nothing). You could
easily look this up without using an AR association if needed and the
call is quite expensive, so I've left that side off for now. That is,
you can't look up what shares a shareable belongs to via a normal
ActiveRecord association right now.

Here's a toy create call to be used in a console to see all of this in action:

```
Share.create!(sharer: User.first,
              shared_miq_templates: [MiqTemplate.first],
              tenants: [Tenant.first],
              miq_product_features: MiqProductFeature.take(10))
```

(transferred from ManageIQ/manageiq@cf3a791d87d15c4963eeb18612b0ad77c45644c1)