Change the invalid migration message to live in a method
The reasoning for this is that RSpec spits out the failing line in
addition to the message.  Since the long string is part of the line, it
appears "twice" in the RSpec message which is confusing.  By moving it
into a method, then the output won't contain the first of the 2
messages.

(transferred from ManageIQ/manageiq@034efaf82cdf94079324f13b35e2312a02eb1bc6)