Migrate oVirt API path from /api to /ovirt-engine/api
Starting with version 4.0 of oVirt, the support for the /api URL path
has been removed, and replaced by /ovirt-engine/api, which was
introduced in oVirt 3.5.  All previous versions of oVirt are already out
of support. The providers that were created with previous versions of
oVirt will have stored in the database the old path, and will stop
working when migrated to oVirt 4.0. To avoid that issue this migration
updates all the relevant providers to use the new supported URL path.

https://bugzilla.redhat.com/1427653

(transferred from ManageIQ/manageiq@009ae63a98119887aed378fd38fb357e82ecff8a)