Cleanup replication spec messages
This prevents the "double rendering" of the long error messages, which
is visually confusing to the developer receiving the error.  The way
RSpec prints errors, it writes the line of code in error to the screen,
followed by the error message.  In these cases, the line of code also
contains the custom error message, so it is rendered twice; first,
uninterpolated as code, and again interpolated as the error.  Moving the
text to a method prevents this.

(transferred from ManageIQ/manageiq@2a371088982be8a1e30fd07d31475e61b2e0aca8)