Fix migration column caching.
When this migration is run in the same process as
20170827091406_change_container_quota_items_columns_to_float.rb, that
migration does some column caching for its ContainerQuotaItem stub model
which ends up being inherited by this migration.  The symptom is that
this migration's stub model does not actually have the created_at
column, and thus never writes the value.  Later on when
change_column_null is called, the column has null values in it causing
that statement to fail.

https://bugzilla.redhat.com/show_bug.cgi?id=1515376