Add status to ext_management_system
Adds a status column to ext_management_system to track when the EMS has
been marked for deletion.  This will allow it to be displayed to the
user as being in the process of being deleted and we can prevent the
user from re-enabling an EMS which is in the process of deletion.