Add switches.host_id to fix modeling/deletions
Currently there is no association between either an EMS or a Host to
their switches that don't go through host_switches.  This means that
switch deletions are "interesting".  This also means that the
manager_ref is not consistent because for shared switches uid_ems is
unique but for host local switches the host is needed.