Move ansible secrets from OpenShift to the database
In previous versions, the secret values for the embedded ansible pod
lived in an OpenShift secret object named "ansible-secrets".In newer
versions these values are managed using the Authentication model
against MiqDatabase.

This migration retrieves the values from the secret and saves them
as "Authentication" objects if we see that we are running in OpenShift
and the secret object is present.

https://www.pivotaltracker.com/story/show/157287338