Use only core libs to get the secret values
Also don't try to remove the other objects during the migration.
We don't want a migration to be tied to a particular gem version
especially if we also will want to change that gem version in the
core app.

We will just have to document any other changes to the project
that need to be made in the course of the upgrade