Add ems_ref to lans to help with unique listing
Add an ems_ref column to the lans table so that it can be used to assist
in getting a unique list of lans across hosts in cases where the uid_ems
is not unique i.e. VMware where the uid_ems is the name of the lan.

By adding the ems_ref which is unique across the ems we are able to get
a truly unique list of networks when showing a list across hosts.