Model relation "SecurityGroup belongs to NetworkRouter or CloudSubnet"
With this commit we add a foreign key on SecurityGroup to
follow what Nuage needs. In Nuage, security group is something
that only exists in a scope of either L3 domain (NetworkRouter) or
L2 domain (CloudSubnet::L2).

MIQ is currently not able to model such relation yet, hence this PR.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>