Combine switches column migrations
This causes an issue with upgrading replicated environments

When the first migration is run, data is generated which references
the new columns in the switches table, but those columns are then removed.

When replication sees data which references a column not in the global
schema it exits and tries again later (typically within a few minutes).
If we then migrate the global database to add then quickly remove the
column in question the data from the remote stream will never be
able to be successfully consumed.

This fixes the issue by not adding the column to the wrong table
to start with.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1601015