replace miq_set relationships with miq_set_memberships table
Note (NickLaMuro): To mirror things properly in the `:down` of the
second migration, it is required to execute 2 queries to add in the
records back into the `relationships` table:

- First:   add in the root nodes for the relationships (parents)
- Second:  use the ids from those records to build the ancestry column
           for the child nodes

This was not done in the previous iteration of this commit, but has been
updated to do so here.

Extra test cases have been added to ensure we are rebuilding the records
properly, and that it only migrates the records we want moved.