Use views for the metrics and metric_rollups tables
This will allow us to use an "INSTEAD OF" trigger to insert the
rows into the sub-tables rather than having to insert and remove
the rows from the base table. This reduces overhead as well as
eliminates table bloat from deleted rows on the base tables.

This approach also preserves the inheritance structure which
continues to allow us to take advantage of the query optomization
from the check constraints on the sub-tables. (This would not
have been possible if we removed the "_base" tables entirely).