Add ansible credentials to authentications
become_username and become_password are used for privilege
escalation on the remote hosts. become_password will be encrypted

auth_key_password is the password to use for decrypting an encrypted
ssh private key. The key itself will be stored in the auth_key
column and both columns will be encrypted

This will allow us to store ansible credentials in a more sane way