Don't sync the schema_migrations_ran table when adding it
The purpose of the table is to track when migrations are run on
remote regions. If we sync the whole table outside of the scope
of the transaction based replication then we're violating the
assumption that the migration version number only appears in the
global region when the data from that migration was replicated
up to the global region.

This was causing a timing issue where the sync could take place
just before the global region runs its migrations. Then the global
would blaze past migrations without successfully replicating up
the data needed for a particular schema version.

This could lead to an irrecoverable situation in replication where
the subscription would have to be removed and re-added.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1721596