Add a migration for moving AWX credentials into our authentications table
This utilizes a python script which is lifted almost directly from
the AWX source to decrypt the extracted fields from the database
ref: https://github.com/ansible/awx/blob/e269634afc9386807ceaf56a61f7f0fad6671bf0/awx/main/utils/encryption.py

We fill in the appropriate attributes in the authentication record
which was previously collected during inventory sync with the
re-encrypted version of the credential field from the awx database.

Attributes which do not fit into our authentication schema are put
into the options serialized column.