Move the embedded ansible proxy settings to a top level git repo key
Without this there is no real way to determine which git repository
instances should pass the proxy settings and which shouldn't.

By making this change we are assuming that if you want a proxy
for getting to your git repositories for embedded ansible you
will likely also want to use that proxy for accessing automate
domain git repos. I think this is a fair assumption.

https://bugzilla.redhat.com/show_bug.cgi?id=1735712