Set zone's visible column default true and update existing rows
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1749694

With hammer (rails 5.0), Zone.seed was not creating the default zone with visible => true, causing it to remain nil.
The hammer UI code wasn't checking this when displaying zones in the UI.

When we upgraded to Ivanchuk/rails 5.1, the UI now checks for visible and these visible => nil zones created in hammer
would not be visible.  Note, Ivanchuk (rails 5.1),  Zone.seed does create the default zone with visible => true.

To be safe, this migration not only relies on postgresql's column default to set Zone visible but also updates
existing records from visible => nil to visible => true.