Handle faulty v2_key while upgrading ansible
Situation:

When importing a foreign database, the v2_key and therefore the
ansible key is not available. Which is good/ok since we shouldn't
be able to read a customer's passwords anyway.

What is bad is there is no way to work with this scenario. The database
simply can not be migrated.

Solution:

Provide a way to continue even though we can't read all the ansible
passwords.

run migration with env variable to just get bogus passwords

INVALID_ANSIBLE_VALUE=bogus rake db:migrate

original PR:
https://github.com/ManageIQ/manageiq-schema/pull/385

https://bugzilla.redhat.com/show_bug.cgi?id=1755553