[MoveAwxCredentialsToAuthentications] Handle faulty v2_key
Situation:

When importing a foreign database, the v2_key and therefore the
ansible key is not available. Which is good/ok since we shouldn't
be able to read a customer's passwords anyway.

What is bad is there is no way to work with this scenario. The database
simply can not be migrated.

Solution:

Provide a way to continue even though we can't read all the ansible
passwords.

run migration with env variable to just get bogus passwords

    $ HARDCODE_ANSIBLE_PASSWORD=bogus rake db:migrate

Also adds some specs around the use cases where the secret_key provided
doesn't exist or does not match what is provided, as well as testing the
workaround in those cases.

original PR:
https://github.com/ManageIQ/manageiq-schema/pull/385

https://bugzilla.redhat.com/show_bug.cgi?id=1755553