[MoveAwxCredentialsToAuthentications] Reset secret_key
Reset the `@secret_key` with each test so we are reloading this with
each spec instead of using the memoized value.

This happened to work previously since we were never changing the
:secret_key value between specs, but that will change in future commits.
This also bumps it outside of the of the `migration_context` block as we
will use a inner block for testing with a different @secret_key value.