Remove ems_ref_obj YAML column
Currently we store a YAML serialized VimString ManagedObjectReference
that includes the vimType (e.g. VirtualMachine/HostSystem/ClusterComputeResource)
because the ems_ref and model are not enough information to accurately
represent a VMware object, you need the VIM class name as well.

Rather than serializing a full VMwareWebService/VimType we can store the
VIM class plus the MOR in a string e.g. 'VirtualMachine:vm-123'. This
gives us enough information to indicate a VMware ManagedObject without
requiring the VMwareWebService gem in core to allow for deserialization
of ems_ref_obj.