Add miq_worker_types table
This table will be seeded to include all leaf subclasses of MiqWorker

On the first pass it will be a replacement for the information currently
stored in the MIQ_WORKER_TYPES constants, but will be extended to
contain all the metadata concerning when workers of each type should
be running (for example, required roles, worker scope, provider type, etc.)

The ultimate goal is to separate out worker management into a standalone
process distinct from the current code-base where a subset of models
will be the primary API between the "worker orchestrator" and the
server which will determine *how* to start the workers it was told
to run.

This will allow for sharing the core logic of when and where to run
workers across different platforms and runtimes such as systemd, fork,
containers, etc.