Split disk_type to 'thin' and 'format' for disks
The thin column will be set to true if the disk_type was 'thin' and
false if it was 'thick'. (In the case of oVirt the 'format' will be set
by the next full refresh)

Otherwise the 'format' will be set to the value of 'disk_type' and the
'thin' value will be set to false (this relevant for the VMware case).

Related to: https://bugzilla.redhat.com/show_bug.cgi?id=1726590