Delete any worker records that have a JSON queue_name
Allowing miq_workers with a compound queue_name is a way to allow a
worker to dequeue for multiple ext_management_systems at once, aka
allowing an Amazon RefreshWorker to refresh the CloudManager and
NetworkManager in the same worker.

This has other side effects however including making it more difficult
to find a worker by an EMS (e.g. when destroying the EMS) and requires
that the base MiqWorker class do `JSON.parse(self[:queue_name]) rescue
self[:queue_name]`

Another option is to have the child managers report their queue_name for
the refresh operation as the parent manager's queue_name.