[db/schema.rb] Support *_trigger functions
In short:  this change moves the `add_trigger` and `drop_trigger`
functionality into the `ActiveRecord::Base.connection` directly (using
some shared conventions from other gems), and also adds support into the
SchemaDumper for database introspection to allow support for triggers
via the `db_schema` (using the same functions now available on the
`.connection` object).

Specifically, two methods needed to be added to support this:

- `ActiveRecord::Base.connection.triggers`
- `ActiveRecord::ConnectionAdapters::SchemaDumper.triggers`

The former provides introspection into the existing database to fetch
the existing triggers that have been created for our database, including
the `direction`, `table`, `name`, and `body`, all of which are the
arguments for our `ActiveRecord::Base.connection.add_trigger` function.

The latter writes the `trigger` statements to the `db/schema.rb` in the
format that the function takes them, formatting the data from the db
into a format that `add_trigger` expects.

* * *

This is a bit dense, but it follows a lot of the conventions for
modifying activerecord that were done in the gem `scenic` to add SQL
view support to postgres:

  https://github.com/scenic-views/scenic/tree/048e0805

So perusing the source code there might help.