[db/schema.rb] Support metric table inheritence
Similar to the additions of `add_trigger` and `drop_trigger`
functionality to the `ActiveRecord::Base.connection` and
`ActiveRecord::ConnectionAdapters::SchemaDumper`, this adds the
following two new methods:

- `add_miq_metric_table_inheritance`
- `drop_miq_metric_table_inheritance`

Since these functions where incredibly "metrics specific", it didn't
seem correct to name them `.add_table_inheritance` as a generic
function, so prefixing with `add_miq...` was how that was handled.

This might make sense for other functions, but left it isolated to this
commit for comparing and contrasting.

* * *

This is a bit dense, but it follows a lot of the conventions for
modifying activerecord that were done in the gem `scenic` to add SQL
view support to postgres:

  https://github.com/scenic-views/scenic/tree/048e0805

So perusing the source code there might help