Introduce Openstack cloud manager child manager migration
Child storage managers were introduced in E release.
When creating a cloud manager, these child managers are auto created.

For existing ems, the cloud volumes were linked to the cloud manager but
they now will point to the storage managers.

to avoid a migration, we had introduced code in the model to do this migration.
This PR moves that code to migrations where it belongs.

This is only necessary for customers transitioning from darga and have
not run a refresh (in the past 5 years...)

Since we are deleting this from the model, we need this in here for safe measure.