Rename TagMapping resource_type for Amazon Vms and Images
The Amazon provider was the first cloud provider to participate in the
(originally) ContainerLabelTagMapping class.  It has since been expanded
to include many providers including Openstack, Azure, and VMware
vSphere.

This means that the original labeled_resource_type of "Vm" and "Image"
meaning Amazon VMs and Templates is quite confusing.