[rails6] Fix create_join_table issues
In Rails 6, a "bug fix" was added to avoid creating duplicate columns:

  https://github.com/rails/rails/pull/33029

This wasn't very well described what this was fixing, but this seems to
be the issue it was fixing:

  https://github.com/rails/rails/issues/33024

However, the problem we had as a result was this error with our specs
when running migrations:

  1) ExpandDialogFieldDefaultValueSize#down with empty tables
     Failure/Error: super

     StandardError:
       An error has occurred, this and all later migrations canceled:

       you can't define an already defined column 'miq_server_id'.
     # ./db/migrate/20160328204930_remove_miq_server_product_update_join_table.rb:21:in `block in down'
     # ./db/migrate/20160328204930_remove_miq_server_product_update_join_table.rb:20:in `down'
     # ./lib/manageiq/schema/migrate_with_cleared_schema_cache.rb:6:in `block in migrate'
     # ./lib/manageiq/schema/migrate_with_cleared_schema_cache.rb:18:in `clearing_caches'
     # ./lib/manageiq/schema/migrate_with_cleared_schema_cache.rb:5:in `migrate'
     # ./spec/support/migration_helper.rb:109:in `block in migrate_to'
     # ./spec/support/migration_helper.rb:99:in `suppress_migration_messages'
     # ./spec/support/migration_helper.rb:105:in `migrate_to'
     # ./spec/support/migration_helper.rb:42:in `prepare_migrate'
     # ./spec/support/migration_helper.rb:8:in `block (2 levels) in migration_context'
     # ------------------
     # --- Caused by: ---
     # ArgumentError:
     #   you can't define an already defined column 'miq_server_id'.
     #   ./db/migrate/20160328204930_remove_miq_server_product_update_join_table.rb:21:in `block in down'

And was expressed by this issue to rails:

  https://github.com/rails/rails/issues/35448

In which the answer was "fix your migrations".

I don't really buy that answer, as that is why they have versioned
migration classes in the first place, but for now, this is the fix
available to us and seems to allow most of the specs to pass.

On top of that, we can't use `create_join_table` as we previously did
since we now can't overwrite default columns and choose what column
type we want to use since the only thing we can provide with the kwargs
is the `:column_options`, but we will always get the foriegn_key columns
as `integer`.

That said, this is a down migration, so probably not worth investing
trying to get this changed upstream.