[rails6][schema_dumper.rb] Use table_name properly
A change that was introduced in Rails 6:

https://github.com/rails/rails/pull/35890/commits/f185e0ebc9cc18c596eb9cab1a9cd99e26a20500

Caused `column.table_name` to be an invalid call, and thus the schema
was not printed out properly.  This wasn't noticed originally because it
was silently passing thanks to this rescue statement:

https://github.com/rails/rails/blob/fe76a95b0d252a2d7c25e69498b720c96b243ea2/activerecord/lib/active_record/schema_dumper.rb#L165-L169

The error that ended up being displayed in the schema was something
like:

  ActiveRecord::Schema.define(version: 2021_01_06_200226) do

    # These are extensions that must be enabled in order to support this database
    enable_extension "plpgsql"

    # Could not dump table "accounts" because of following NoMethodError #   undefined method `table_name' for
    #<ActiveRecord::ConnectionAdapters::PostgreSQL::Column:0x000055ae1dbe7d70>

In the same Rails commit that removed the `column.table_name`, a
`table_name` attribute was added as an attr_accessor to `SchemaDumper`
to track that current `table_name` being worked on, so that was used in
this fix instead of the previous `table_name` method on `Column`.

* * *

This commit also adds schema_dumper_spec.rb, which is meant to catch
issues like this in the future in CI instead of by another user.