Add a JSONB Capabilities to EMS and Providers
Allow storing provider and manager capabilities in VMDB to assist in
deciding what is possible to do on a particular provider instance
without having to do an API call out to the provider.

This is commonly done to decide if we can e.g. start an event catcher or
collect metrics.