More accurately check schema dumper lines
Prior to this commit, the code would check that the line in the
schema.rb by using a substring match, which could be inaccurate,
particularly when a table name may be a substring of another table name.
This commit changes the lookups to only look at the full table names.