Drop unused pglogical references
As of https://github.com/ManageIQ/manageiq/pull/18686 we moved from pglogical
to the pg-logical_replication using postgresql's native logical replication.

We switched as of ivanchuk.  It should be safe to drop this now.

Funny enough, this actually helps with getting ruby 3.0 support because the AR
extension signature for drop_table needs to change to work with ruby 2.x and
3.x.