Add ExtManagementSystem#last_refresh_success_date
Currently we have `:last_refresh_date` and `:last_refresh_error`, but if
the refresh is currently failing we don't have a way to see when the
last successful refresh was.

Adding a `:last_refresh_success_date` allows us to show users when
refresh was working if it is currently failing.