move concurrent requests out of global scope
2 settings concepts have been merged

Had originally intended to keep the batch_size, but there wasn't a good way to keep it
and avoid duplicates from the /performance/concurrent_requests keys.
It is so new that we are ignoring it for now. No one would have populated the value.


The values that are currently used are 20,1,1 -- which are way low.
locally I use 250,100,100 -- so may want to keep that in mind