Remove VimStrings from MiqQueue
Events used to have vmware specific objects in the queue.
These leaked through for a customer and it blew up.

This has not been in the queue for a number of years, but if someone delays
upgrading and then doesn't empty the queue before upgrading... this will happen.

The solution is to remove entries