Set manager_ref to file path for embedded scripts
The manager_ref is typically used to track the "unique" attribute of an
object, for ansible_tower/awx the manager_ref is the job_template.id but
for embedded_ansible the manager_ref was nil and the name was set to the
file path.  Embedded Workflows copied this behavior, but it would be
preferable to match how the "external" automation managers operate and
also open up the `:name` column for a more user friendly name.