Fix OrchestrationStack#configuration_script reference
There are some `OrchestrationStack` types (AWX and AnsibleTower Jobs)
which were using `orchestration_template_id` to reference a
configuration_script.  This was causing issues with the other side of
the relationship, namely if there was an `OrchestrationTemplate` with the
same ID as the `ConfigurationScript` then that template would think it had
active stacks and could not be deleted.

There is already a `OrchestrationStack#configuration_script_base_id` but
that is used to track the `Playbook` aka the `ConfigurationScriptPayload`
where this is used to track the `JobTemplate` which is a running
`ConfigurationScript`.