Allow the rails method to be called, don't expect it
In rails 7.0, it's called 10 times plus 1 more on the next line.  In rails 7.1+,
it's called only once.  From a test perspective, we don't care if it's even called.
We only care that the SQL query is executed with the scramsha password or not.

To change the username/password only in calls from the migration itself and not from
rails, we hack around the caller_locations to make the change only for the migration callers.

With this, I believe we can expect test success on rails 7.1 and 7.2.

It was added in 742841cf8954be1cb5843dca33945b5fc001cf8d but I think this change
emphasizes what we really care about testing and removes an assertion on internals
that we don't really have interest in testing.

Fixes #768