Split inherit/constraint for usage in collapsed migration/schema.rb
We need to do inheritance in both migrations and from schema dump/load.

For constraints, we don't need to do them from schema load as t.check_constraints
is now included for each metric_* subtable in the dumped schema.rb, so we no
longer pass the conditions from schema.rb dump/load.

Simplify the interface to add_miq_metric_table_constraint to pass the array of
conditions inline.

Note, this issue was found in #824 and extracted and simplified here.