8.0 supports inherits in schema dump but doesn't handle table order
This commit allows rails 7.2 to operate in the same way as 8.0.
We provide a copy of a rails 8.0 method, inherited_table_names, allowing
us to use quite similar code in 7.2 and 8.0.

We support dumping subtables in 7.2 and 8.0 by following the same pattern:
1) Dump the base tables by ignoring the subtables
2) Dump the subtables by ignoring the base tables

This is because metric_rollups_base is sorted alphabetically AFTER
metric_rollups_01, etc. The same problem happens with metrics_base sorting
AFTER metrics_00.

Rails 8.0 added inherits support in schema dumper here:
https://www.github.com/rails/rails/pull/50475

A possibly related issue regarding table order with dumping inherits:
https://www.github.com/rails/rails/issues/54841

A possible solution may come in a future release as part of:
https://www.github.com/rails/rails/pull/54872

Tracking was a rails 7.2 method that is no longer needed with the backport of
the 8.0 method, inherited_table_names.